% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runShinyApp.R
\name{runShinyApp}
\alias{runShinyApp}
\title{The shiny application \code{AFheritability} is a user interface for the function \code{\link[AFheritability]{AFfunction}}}
\usage{
runShinyApp()
}
\description{
The shiny-app provides a user friendly interface for the function \code{\link[AFheritability]{AFfunction}}.
}
\details{
By running runShinyApp() a user interface for the function \code{\link[AFheritability]{AFfunction}} is started in RStudio. The app is also available online \url{https://afheritability.shinyapps.io/afheritability/} (Note that the app is usually faster in the web browser Google Chrome or Firefox).
}
\references{
Dahlqwist E et al. (2019) <doi:10.1007/s00439-019-02006-8>.
}
\author{
Elisabeth Dahlqwist
}
